+++ src/commands/admin/follow.js
@@ -20,7 +20,7 @@ module.exports = class CMD extends Command {
                 },
                 disabled: false,
                 cooldown: 10,
-                category: "Setup",
+                category: "Administration",
             },
             client
         );