+++ src/commands/admin/modlogs.js
@@ -24,7 +24,7 @@ module.exports = class CMD extends Command {
                 ],
                 disabled: false,
                 cooldown: 10,
-                category: "Setup",
+                category: "Administration",
             },
             client
         );